wrap call to os.getcwd() inside try/catch block.
within an interactive python session calling this method seems to throw
an "OSError: [Errno 2] No such file or directory" error. wrap it in the
try/except to silence the error and continue onward. This should only
affect loading of config files in the "current working directory" when
in an interactive session (a rare occurance), and there are several
other methods of injecting config files (e.x. environment var, etc).