Added two new abstract methods to PipelineModule:
__declare_parameters() - all parameter declerations should be performed
in this method.

__declare_resolvers() - all resolver declerations should be performed in
this method.

This change is important so that the parent class can automatically call
the set_parameters_from_config() method after parameter decleration.
Otherwise child classes were responsible for calling the load config
method, something that was not being done consistently. This should help
force the issue.

This commit also includes changes to the moduels themselves that
implement the above changes. (exception with MACS2, where more of the
MACS2 command line options were implemented as module parameters).